chore: bridge polling duration
